+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -221,5 +221,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-25T12:37:14.599Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-26T07:53:12.733Z")
     r"""The date and time the resource was last updated"""