+++ packages/orq-rc/RELEASES.md
@@ -909,3 +909,13 @@ Based on:
 - [python v3.3.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.9] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.9 - packages/orq-rc
+
+## 2025-03-26 07:54:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.523.2 (2.560.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.10] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.10 - packages/orq-rc